Refactored routes and controllers, cleaned up index file for better structure
